edit: update file path, scroll detail animation
